Fix Request/Response Flow Blocked on Bad Command Written.
The byte array that we were writing was passed as &byte. We were
therefore writing the first 7 bytes of the pointer of that byte array.
Yes, it has been a while since I wrote C but I remember some of it.

Now that the bytes that are written are the ones that we're meaning to
send, the response is actually received back from the device.