Implement Alternative Method of Resolving Time.
* Resolve timestamps using the offset for the current system time vs the
  current timestamp on the host.
* Resolve timezone offset from the offset between the display time /
  event time and the internal time. All time values are adjusted with
  the offset described above.

This should be more robust than assuming that the internal system time
of the dexcom receivers are standard. Also, the display time offsets is
affected by the same flaw. The dexcom display time offset is also
relative to the system time. Since the system time does not appear to be
standard, it does not help by itself.