Simplify Timezone Handling by Using Local TimeZone.
Instead of trying to detect the timezone from the dexcom, the
problem has been simplified. The timezone now comes from the
local system. Therefore, it's more important to have timestamps
correct.

The internal timestamp is pretty safe since we have
the current system time from the dexcom from which we extract
the offset with the real epoch. However, display time is tricky
because it still represents an offset that matches the user time
and it's still the same instant in the time continum. This will
need revisiting at some point. Likely, the display time should
just go away and we would only keep the display time as an
offset to avoid confusion.

The event time is better than the display time though. Its time
value should be a real timestamp. The way this is done is that
we discount the event time value by the offset from the raw
display time against the system time. This should give us
the real time of the event.